+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.13.0 (2016-06-11)
+Features:
+  - Double-click to add spaces.
+  - Use "justify" as the default align.
+  - Also squeeze the first CJK quotation marks.
+
+Bugfixes:
+  - Fix a bug in fixing "reverse-paired" marks. The function is too buggy and disabled by default.
+  - Blank spaces at the beginning of a line is not deleted anymore.
+  - Elements without text nodes can also be fixed.
+
 ## 0.12.5 (2016-06-10)
 Features:
   - Use the asynchronous feature to fire re-checking after other events are done.